one to many relationship

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
+import { User } from './../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,12 +15,15 @@ export class TasksService {
 		private taskRepository: TaskRepository
 	) {}
 
-	async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-		return this.taskRepository.getTasks(filterDto);
+	async getTasks(
+		filterDto: GetTasksFilterDto,
+		user: User
+	): Promise<Task[]> {
+		return this.taskRepository.getTasks(filterDto, user);
 	}
 
-	async getTaskById(id: number): Promise<Task> {
-		const found = await this.taskRepository.findOne(id);
+	async getTaskById(id: number, user: User): Promise<Task> {
+		const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 		
 		if (!found) {
 			throw new NotFoundException(`Task with ${id} not found`);
@@ -28,22 +32,30 @@ export class TasksService {
 		return found;
 	}
 
-	async createTask(createTaskDTO: CreateTaskDto): Promise<Task> {
-		return this.taskRepository.createTask(createTaskDTO);
+	async createTask(
+		createTaskDTO: CreateTaskDto,
+		user: User	
+	): Promise<Task> {
+		return this.taskRepository.createTask(createTaskDTO, user);
 	}
 
-	async deleteTask(id: number): Promise<void> {
-		const result = await this.taskRepository.delete(id);
+	async deleteTask(
+		id: number,
+		user: User
+	): Promise<void> {
+		const result = await this.taskRepository.delete({ id, userId: user.id });
 
 		if (result.affected === 0) {
 			throw new NotFoundException(`Task with ID ${id} not found`);
 		}
 	}
 
-	async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-		const task = await this.getTaskById(id);
-
-		console.log(task);
+	async updateTaskStatus(
+		id: number,
+		status: TaskStatus,
+		user: User
+	): Promise<Task> {
+		const task = await this.getTaskById(id, user);
 
 		task.status = status;
 		await task.save();
